+++ src/frontend/komponenter/Fagsak/BehandlingContainer.tsx
@@ -14,6 +14,8 @@ import {
     finnSideAktivtSteg,
 } from '../Felleskomponenter/Venstremeny/sider';
 import Venstremeny from '../Felleskomponenter/Venstremeny/Venstremeny';
+import BrevmottakerContainer from './Brevmottaker/BrevmottakerContainer';
+import { BrevmottakerProvider } from './Brevmottaker/BrevmottakerContext';
 import FaktaContainer from './Fakta/FaktaContainer';
 import { FeilutbetalingFaktaProvider } from './Fakta/FeilutbetalingFaktaContext';
 import { FeilutbetalingForeldelseProvider } from './Foreldelse/FeilutbetalingForeldelseContext';
@@ -21,8 +23,6 @@ import ForeldelseContainer from './Foreldelse/ForeldelseContainer';
 import Høyremeny from './Høyremeny/Høyremeny';
 import { FeilutbetalingVedtakProvider } from './Vedtak/FeilutbetalingVedtakContext';
 import VedtakContainer from './Vedtak/VedtakContainer';
-import VergeContainer from './Verge/VergeContainer';
-import { VergeProvider } from './Verge/VergeContext';
 import { FeilutbetalingVilkårsvurderingProvider } from './Vilkårsvurdering/FeilutbetalingVilkårsvurderingContext';
 import VilkårsvurderingContainer from './Vilkårsvurdering/VilkårsvurderingContainer';
 
@@ -158,9 +158,9 @@ const BehandlingContainer: React.FC<IProps> = ({ fagsak, behandling }) => {
                     <Route
                         path={BEHANDLING_KONTEKST_PATH + '/verge'}
                         element={
-                            <VergeProvider behandling={behandling} fagsak={fagsak}>
-                                <VergeContainer />
-                            </VergeProvider>
+                            <BrevmottakerProvider behandling={behandling} fagsak={fagsak}>
+                                <BrevmottakerContainer />
+                            </BrevmottakerProvider>
                         }
                     />
                 </Routes>
